+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Reminders
 
+## Unreleased
+
+- Added a "What's New" window that shows the changelog in-game: it opens automatically the first time you log in after updating, and there's a "What's New" button on the Reminders window to reopen it any time (#55)
+
 ## [v12.4.0](https://github.com/pcg79/Reminders-Addon/tree/v12.4.0) (2026-07-27)
 [Full Changelog](https://github.com/pcg79/Reminders-Addon/compare/v12.3.0...v12.4.0)